move proto_cmd.py into the setup.py, hard-code version so that sdist can work as expected.
